+++ .github/tests/application/tests/TestCases/PluginIsActive/ForumsNumericTest.php
@@ -23,7 +23,7 @@ class ForumsNumericTest extends AbstractForumsTest
         $this->forumsAreEmpty = false;
     }
 
-    #[Attributes\DependsOnClass(ForumsPagedTest::class)]
+    #[Attributes\DependsOnClass(ForumPagedTest::class)]
     #[Attributes\DataProviderExternal(ForumDataProvider::class, 'getForums')]
     public function testForumAsGuest(Forum $forum): void
     {
