+++ .github/tests/application/tests/TestCases/PluginIsActive/TopicNumericPagedTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\PluginIsNotActive;
+namespace Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\TestCases\PluginIsActive;
 
-use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Abstracts\AbstractTopicsPagedTest;
+use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Abstracts\AbstractTopicPagedTest;
 use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\DataProviders\ForumDataProvider;
 use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Entities\Posts\Forum;
 use Korobochkin\BBPressPermalinksWithIdTestsApplication\Tests\Entities\Posts\Topic;
@@ -14,9 +14,16 @@ use PHPUnit\Framework\Attributes;
  * @internal
  */
 #[Attributes\CoversNothing]
-class TopicsPagedTest extends AbstractTopicsPagedTest
+class TopicNumericPagedTest extends AbstractTopicPagedTest
 {
-    #[Attributes\DependsOnClass(RepliesTest::class)]
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->useNumericPermalinksRequests = true;
+        $this->useNumericPermalinksHTML = true;
+    }
+
+    #[Attributes\DependsOnClass(TopicsNumericTest::class)]
     #[Attributes\DataProviderExternal(ForumDataProvider::class, 'getRepliesPaged')]
     public function testTopicPagedAsGuest(Forum $forum, Topic $topic, int $page, array $replies): void
     {
