+++ .github/tests/application/tests/TestCases/PluginIsNotActive/RepliesTest.php
@@ -37,7 +37,7 @@ class RepliesTest extends AbstractRepliesTest
         }
     }
 
-    #[Attributes\DependsOnClass(ForumsPagedTest::class)]
+    #[Attributes\DependsOnClass(ForumPagedTest::class)]
     #[Attributes\DataProviderExternal(ForumDataProvider::class, 'getReplies')]
     public function testCreateReply(Forum $forum, Topic $topic, Reply $reply): void
     {
